DonChamba WebApp v1

+++ Web_App_DonChamba/Models/Producto.cs
@@ -14,7 +14,7 @@ namespace WebAPIDonChamba.Models
         public int FkIdCategoria { get; set; }
         public string Nombre { get; set; } 
         public string Descripcion { get; set; }
-        public byte Estado { get; set; }
+        public int Estado { get; set; }
         public string Imagen { get; set; } 
         public decimal Precio { get; set; }
